task - 12 figures

+++ src/main/java/pl/coderstrust/Figures/Figure.java
@@ -1,5 +1,5 @@
 package pl.coderstrust.Figures;
 
 public interface Figure {
-    long calculateArea();
+    double calculateArea();
 }
